GlobalHandlerException, logic for trx, agreement and account

+++ src/main/java/com/example/bankapp/entities/AgreementEntity.java
@@ -6,6 +6,7 @@ import lombok.Setter;
 import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.UpdateTimestamp;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 import java.util.Date;
 
@@ -26,7 +27,7 @@ public class AgreementEntity {
     @JoinColumn(name = "account_id")
     private AccountEntity account;
 
-    //@Column(name = "account_id")
+   // @Column(name = "account_id")
     //private Long accountId;
 
     @ManyToOne
@@ -40,17 +41,14 @@ public class AgreementEntity {
     @JoinColumn(name = "manager_id")
     private ManagerEntity manager;
 
-    //@Column(name = "manager_id")
-    //private Long managerId;
-
     @Column(name = "interest_rate")
-    private int interestRate;
+    private double interestRate;
 
     @Column(name = "status")
     private int status;
 
     @Column(name = "sum")
-    private double sum;
+    private BigDecimal sum;
 
     @CreationTimestamp
     @Column(name = "created_at")
